CAP-79 #comment Organize files
