+++ static/swagger.yml
@@ -40,6 +40,8 @@ paths:
       responses:
         200:
           description: Contents of table
+        404:
+          description: Table not found
         500:
           $ref: "#/responses/Standard500ErrorResponse"
